asm: extract to separate module
