+++ README.md
@@ -3,7 +3,7 @@ Learning purpose riscv64 (RV64GC) emulator.
 This project is built for [一生一芯](https://ysyx.org/) as a reference implementation.
 
 ### Highlights
-- [Type-safe instructions](./valheim-core/src/isa/typed.rs) which makes the decoding [less error-prone](./valheim-core/src/isa/decode.rs).
+- [Type-safe instructions](valheim-asm/src/isa/typed.rs) which makes the decoding [less error-prone](valheim-asm/src/isa/decode.rs).
 - Full emulation trace (registers, memory, etc.) like persistent data structures, which is useful for debugging the real hardware.
 - [MISA]() = `RV64ACDFIMSU`
   - RV64G (IMAFD_Zicsr_Zifencei) instruction set
