+++ valheim-asm/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "valheim-asm"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+modular-bitfield = "0.11.2"
