+++ valheim-asm/src/asm/encode16.rs
