+++ valheim-asm/src/isa/typed.rs
