+++ valheim-asm/src/lib.rs
@@ -1,0 +1,8 @@
+#![allow(incomplete_features)]
+
+#![feature(core_intrinsics)]
+#![feature(generic_const_exprs)]
+#![feature(inline_const_pat)]
+
+pub mod asm;
+pub mod isa;
