+++ valheim-core/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-modular-bitfield = "0.11.2"
+valheim-asm = { path = "../valheim-asm" }
 memmap2 = "0.5.3"
 derive_more = "0.99.0"
 rustc_apfloat = "0.1.3"
