+++ valheim-core/src/cpu/irq.rs
@@ -1,5 +1,5 @@
 use crate::cpu::{PrivilegeMode, RV64Cpu};
-use crate::cpu::csr::CSRMap::{MCAUSE, MEDELEG, MEIP_MASK, MEPC, MIDELEG, MIE, MIP, MSIP_MASK, MTIP_MASK, MTVAL, MTVEC, SCAUSE, SEIP_MASK, SEPC, SIE, SIP, SSIP_MASK, SSTATUS, STIP_MASK, STVAL, STVEC};
+use crate::cpu::csr::CSRMap::{MCAUSE, MEDELEG, MEIP_MASK, MEPC, MIDELEG, MIE, MIP, MSIP_MASK, MTIP_MASK, MTVAL, MTVEC, SCAUSE, SEIP_MASK, SEPC, SSIP_MASK, STIP_MASK, STVAL, STVEC};
 use crate::device::virtio::Virtio;
 use crate::memory::VirtAddr;
