+++ valheim-core/src/lib.rs
@@ -7,9 +7,7 @@
 extern crate core;
 #[macro_use]
 extern crate derive_more;
-pub mod asm;
 pub mod cpu;
-pub mod isa;
 pub mod memory;
 pub mod interp;
 pub mod device;
